requirements.txt
translatepy
openai
tenacity
ruamel.yaml